refactor: PlaceInfo와 UserVisitPlace의 패키지 위치 이동
